Switch to `CStr::from_bytes_until_nul` on sized char arrays in structs
Certain structs contain sized character arrays that are converted to
`CStr` for debugging using unsafe `CStr::from_ptr(...as_ptr())`.  There
is no need to round-trip to a pointer and possibly read out of bounds if
the NULL character index (string length) is instead scanned for manually
on the slice or via the newly stabilized `CStr::from_bytes_until_nul()`
since Rust 1.69 (which panics if no NULL char is found before the end of
the slice).

Unfortunately `unsafe` is still needed to cast the array from a `c_char`
(`i8` on most platforms) to `u8`, which is what `from_bytes_until_nul()`
accepts.